judge/requirements.txt
flask
flask-restful
pip-chill
pandas
umap-learn